backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.9.0.80
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
python-srt==3.5.0
